requirements.txt
python-dotenv
pandas 
numpy 
statsmodels
pymongo
Certifi
pymongo[srv]==3.6
scikit-learn
PyYAML




